Move getresumestate to FormLanding Component
