+++ src/components/form-landing/form-landing.tsx
@@ -1,15 +1,31 @@
 import React, { useState, useEffect } from 'react';
 import { Vedtak } from '../../types/vedtak';
 import { useDispatch, useSelector } from 'react-redux';
-import { checkAuth, getExistingKlage } from '../../store/actions';
+import {
+    checkAuth,
+    getExistingKlage,
+    postNewKlage,
+    setKlageId,
+    setStorageContent,
+    setValgtTema,
+    setValgtYtelse
+} from '../../store/actions';
 import { Store } from '../../store/reducer';
 import WithLoading from '../../components/general/loading/withLoading';
-import { logInfo } from '../../utils/logger/frontendLogger';
+import { logError, logInfo } from '../../utils/logger/frontendLogger';
 import { queryToVedtak } from '../../mock-api/get/vedtak';
 import MainFormPage from '../../pages/form-landing-page/main-form-page';
 import Error from '../../components/error/error';
 import queryString from 'query-string';
 import * as H from 'history';
+import { AxiosError } from 'axios';
+import { getReferrer, getTemaObject } from '../../services/klageService';
+import { KlageSkjema } from '../../types/klage';
+import { getResumeState } from '../../utils/get-resume-state';
+import { DatoValg } from '../begrunnelse/datoValg';
+import NavFrontendSpinner from 'nav-frontend-spinner';
+import { CenteredContainer } from '../../styled-components/main-styled-components';
+import NotFoundPage from '../../pages/not-found/not-found-page';
 
 interface Props {
     query: queryString.ParsedQuery<string>;
@@ -25,10 +41,67 @@ const FormLanding = (props: Props) => {
 
     const [chosenVedtak, setChosenVedtak] = useState<Vedtak>();
     const [temaNotSet, setTemaNotSet] = useState<boolean>(false);
+    const [isLoading, setIsLoading] = useState<boolean>(true);
+    const [errorState, setErrorState] = useState<boolean>(false);
 
     useEffect(() => {
         dispatch(checkAuth(props.location.search));
-        if (typeof klageId !== 'undefined' && klageId.length !== 0 && typeof activeKlage === 'undefined') {
+
+        const { klageId, tema, ytelse, saksnummer } = getResumeState(
+            props.location.search,
+            sessionStorage,
+            props.location.pathname
+        );
+        setStorageContent(klageId, tema, ytelse, saksnummer);
+
+        if (ytelse !== null) {
+            dispatch(setValgtYtelse(ytelse));
+        }
+        if (tema !== null) {
+            dispatch(setValgtTema(tema));
+        }
+
+        if (klageId !== null) {
+            dispatch(setKlageId(klageId));
+            setIsLoading(false);
+        } else if (ytelse !== null && tema !== null) {
+            const klageSkjema: KlageSkjema = {
+                id: null,
+                ytelse,
+                tema,
+                saksnummer: saksnummer,
+                datoalternativ: DatoValg.INGEN,
+                vedtak: null,
+                fritekst: '',
+                vedlegg: [],
+                referrer: getReferrer()
+            };
+            dispatch(postNewKlage(klageSkjema));
+            setIsLoading(false);
+        } else if (tema !== null) {
+            getTemaObject(tema)
+                .then(temaObject => {
+                    dispatch(setValgtYtelse(ytelse ?? temaObject.value));
+                    setIsLoading(false);
+                })
+                .catch((err: AxiosError) => {
+                    if (err.response?.status === 404) {
+                        setErrorState(true);
+                        setIsLoading(false);
+                        return;
+                    }
+                    logError(err);
+                });
+        } else {
+            setIsLoading(false);
+        }
+
+        if (
+            typeof klageId !== 'undefined' &&
+            klageId !== null &&
+            klageId.length !== 0 &&
+            typeof activeKlage === 'undefined'
+        ) {
             dispatch(getExistingKlage(klageId));
         }
 
@@ -38,9 +111,21 @@ const FormLanding = (props: Props) => {
         }
 
         setTemaNotSet(chosenTema === null);
-    }, [dispatch, props.location.search, props.query, chosenTema, klageId, activeKlage]);
+    }, [dispatch, props.location.search, props.location.pathname, props.query, chosenTema, klageId, activeKlage]);
 
     logInfo('Form landing page visited.', { chosenYtelse: chosenYtelse, referrer: document.referrer });
+
+    if (isLoading) {
+        return (
+            <CenteredContainer>
+                <NavFrontendSpinner type={'XL'} />
+            </CenteredContainer>
+        );
+    }
+
+    if (errorState) {
+        return <NotFoundPage />;
+    }
 
     if (temaNotSet) {
         logInfo('Form landing page visited with no tema.', { referrer: document.referrer });
